drivers: fix errors with newer clang

+++ drivers/misc/mediatek/freqhopping/fhctl_new/clk-fhctl-mcupm.c
@@ -305,7 +305,7 @@ static int mcupm_init_v1(struct pll_dts *array, struct match *match)
 	match_data = match->hdlr->data;
 	if (match_data && match_data->reg_tr) {
 		priv_data->reg_tr = array->fhctl_base
-			+ (unsigned int)match_data->reg_tr;
+			+ (unsigned int)(long)match_data->reg_tr;
 	}
 
 	/* hook to array */
